checkboxes working fine
data is rerendering, not puking

it is okay that the shape of the plot never changes, because all of the
values are constant over time, and the deviations from constant are
exactly based on the factors

for a single grid cell *with* detector data, the factors are based on
that grid's detector data.  For a grid cell without detector data, the
factors are based on the model.  Regardless, the factors are multiplying
a constant, which is added up, so taking out one or more of those
constant values only shifts the scale, not the shape.